Attached the arrows to the robot
Removed old physics junk
Modified model to be centered at its CoM

+++ Simulator/Serialization/RobotConfig.cs
@@ -13,6 +13,6 @@ namespace Simulator.Serialization
 
         public Point3D CenterOfMass { get; set; }
 
-        public int Mass { get; set; }
+        public double Mass { get; set; }
     }
 }
